+++ src/Soluble/Japha/Bridge/Driver/Pjb62/PjbProxyClient.php
@@ -231,7 +231,7 @@ class PjbProxyClient
      * @param mixed $args Arguments to send to method
      * @return mixed
      */
-    public function invokeMethod($object, $method, array $args=[])
+    public function invokeMethod(JavaType $object, $method, array $args=[])
     {
         $id = ($object == null) ? 0 : $object->__java;
         return self::$client->invokeMethod($id, $method, $args);