+++ src/Soluble/Japha/Bridge/Driver/Pjb62/Pjb62Driver.php
@@ -123,9 +123,8 @@ class Pjb62Driver extends AbstractDriver
         return $this->pjbProxyClient->getClient()->getContext();
     }
 
-
     /**
-     *
+     * Inspect the class internals
      *
      * @param Interfaces\JavaObject $javaObject
      * @return string